+++ perl/LinkedSpec/ActionIR/Scanner/LegacyRules.pm
@@ -88,6 +88,8 @@ sub try_scan_contract_ir_events {
   'cursor_col' => \&_scan_contract_cursor_col,
   'cursor_rest' => \&_scan_contract_cursor_rest,
   'cursor_rest_len' => \&_scan_contract_cursor_rest_len,
+  'input_text' => \&_scan_contract_input_text,
+  'input_len' => \&_scan_contract_input_len,
   'entry_text' => \&_scan_contract_entry_text,
   'entry_group' => \&_scan_contract_entry_group,
   'entry_groups' => \&_scan_contract_entry_groups,
@@ -765,6 +767,24 @@ sub _scan_contract_cursor_rest_len {
  my ($code) = @_;
  my @events;
 while ($code =~ /\bcursor_rest_len\s*\(\s*\)/g) {
+ push @events, {raw => $&, args => {}};
+}
+ return \@events
+}
+
+sub _scan_contract_input_text {
+ my ($code) = @_;
+ my @events;
+while ($code =~ /\binput_text\s*\(\s*\)/g) {
+ push @events, {raw => $&, args => {}};
+}
+ return \@events
+}
+
+sub _scan_contract_input_len {
+ my ($code) = @_;
+ my @events;
+while ($code =~ /\binput_len\s*\(\s*\)/g) {
  push @events, {raw => $&, args => {}};
 }
  return \@events